+++ src/main/java/service/DepositOperations.java
@@ -3,5 +3,5 @@ package service;
 import model.BankAccount;
 
 public interface DepositOperations {
-    void deposit(BankAccount account, int amount);
+    void deposit(BankAccount account, double amount);
 }
